requirements.txt
# Paquetes principales
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
pandas>=1.5.0
tqdm>=4.0.0
together>=1.0.0
Pillow>=9.0.0
pillow-heif>=0.10.0

# Dependencias específicas para manejo de archivos
python-dotenv>=0.19.0  # Para manejo de variables de entorno

# Dependencias opcionales (para funcionalidades avanzadas)
psutil>=5.8.0  # Para monitoreo de procesos (solución de bloqueo de archivos)
pywin32>=300 ; sys_platform == 'win32'  # Solo para Windows (manejo avanzado de archivos)